Create LABPhotoGalleryTableHTML
The lab contains an HTML file, three image files, and a CSS file. Click the Files icon (with a 5), located in the code editor's top-left corner, to see the lab's five files. The index.html file uses a <link> element to import the CSS file styles.css. The CSS changes the font, adds a border around the table, and colors every other row light gray.

The table should have a <caption> element that describes the photo gallery's theme. The table should have 3 columns and 5 rows:

The 1st row should use <th> elements to list the Photo, Date, and Description headings.

The 2nd - 4th rows should display the given vacation images in the 1st column with an alt attribute that briefly describes the photo. The 2nd column should display the date when the photo was taken. The 3rd column should have a lengthier photo description.

The 5th row should span all 3 columns and contain copyright information, including the copyright symbol.

Except the for the three column headings, your solution does not have to use the exact same text as the example above.